refactor: replace `lib` with `@boxfish-studio/candymachine-client-sdk` (#223)
* feat: add lib and start replacing

* feat: replace lib for pkg and update anchor

* feat: move enum

* feat: remove lib

* fix: fix cacheData type

* fix: fix notification style

* fix: fix double call to fetchNfts
