--- lib/candy-machine/enums.ts
@@ -1,16 +1,0 @@
-export enum StorageType {
-    ArweaveBundle = 'arweave-bundle',
-    ArweaveSol = 'arweave-sol',
-    Arweave = 'arweave',
-    Ipfs = 'ipfs',
-    Aws = 'aws',
-    NftStorage = 'nft-storage',
-    Pinata = 'pinata',
-}
-
-export enum CandyMachineAction {
-    Create = 'creating',
-    Update = 'updating',
-    Delete = 'deleting',
-    Fetch = 'fetching',
-}
