--- lib/candy-machine/index.ts
@@ -1,2 +1,0 @@
-export * from './update'
-export * from './upload'
