--- lib/candy-machine/types.ts
@@ -1,1 +1,0 @@
-export type Account = string | string[] | undefined
