--- lib/candy-machine/upload/index.ts
@@ -1,5 +1,0 @@
-export * from './arweave'
-export * from './config'
-export * from './helpers'
-export * from './transactions'
-export * from './upload'
