+++ lib/enums.ts
@@ -1,0 +1,6 @@
+export enum CandyMachineAction {
+    Create = 'creating',
+    Update = 'updating',
+    Delete = 'deleting',
+    Fetch = 'fetching',
+}
