+++ styles/globals.scss
@@ -94,7 +94,7 @@ body {
 
 .notifications {
     z-index: 100;
-    position: absolute;
+    position: fixed;
     bottom: 0px;
     right: 0px;
     width: 100%;
